Calculate IP prefix for masking only once.
This avoids some instructions in the inner loop since the prefix is
typically unchanged during the run-time. If the mask will be changed
during run-time (e.g. as done in the tests), the prefix will be
updated accordingly.

Also add setters and getters for ipv4mask and ipv6mask for
compatibility and respective test-cases.