Use a dict to store and access the IP prefixes used for masking.
This avoids some more instructions in the inner loop, since
the prefix accessed by simply using the IP version as dict-key.

Test showed that using a dict for storing the prefixes is at least as
fast as using a list. Thus a dict was choosen as it is more elegant.